+++ packages/x-adapter-platform/src/mappers/__tests__/url.spec.ts
@@ -1,4 +1,4 @@
-import { extractUrlParameters, getTaggingInfoFromUrl } from '../url.utils';
+import { extractUrlParameters, getDisplayClickTagging, getTaggingInfoFromUrl } from '../url.utils';
 
 describe('url utils methods tests', () => {
   describe('extractUrlParameters', () => {
@@ -36,6 +36,28 @@ describe('url utils methods tests', () => {
       expect(url).toBe('https://api.empathy.co/');
       expect(params).toStrictEqual({
         q: 'chips',
+        env: 'mobile',
+        lang: ['english', 'spanish'],
+        follow: false
+      });
+    });
+  });
+
+  describe('getDisplayClickTagging', () => {
+    it('should not break when dealing with bad urls', () => {
+      expect(getDisplayClickTagging('null')).toStrictEqual({
+        url: 'null',
+        params: { displayId: undefined, follow: false }
+      });
+    });
+
+    it('should retrieve the tagging info from the url, replacing q with displayId', () => {
+      const { url, params } = getDisplayClickTagging(
+        'https://api.empathy.co/?q=chips&env=mobile&lang=english&lang=spanish'
+      );
+      expect(url).toBe('https://api.empathy.co/');
+      expect(params).toStrictEqual({
+        displayId: 'chips',
         env: 'mobile',
         lang: ['english', 'spanish'],
         follow: false